third_party/WebKit/LayoutTests/platform/win/bluetooth/descriptor/writeValue/gen-io-op-fails-expected.txt

CONSOLE ERROR: Uncaught (in promise) NotFoundError: No Descriptors matching UUID 000000a1-97e5-4cd7-b9f1-f5a427670c59 found in Characteristic with UUID 000000a1-97e5-4cd7-b9f1-f5a427670c59.
This is a testharness.js-based test.
PASS writeValue fails. Should reject with appropriate error. 
Harness: the test ran to completion.
